Refactored the views into a nicely structured module
